+++ lib/firebase_cloud_messenger/webpush/config.rb
@@ -1,7 +1,7 @@
 module FirebaseCloudMessenger
   module Webpush
     class Config < FirebaseObject
-      FIELDS = %i(headers data notification).freeze
+      FIELDS = %i(headers data notification fcm_options).freeze
       attr_accessor(*FIELDS)
 
       def initialize(data)